move to RscTitles, handle vehicle
