+++ addons/medical_feedback/config.cpp
@@ -19,4 +19,3 @@ class CfgPatches {
 #include "CfgSounds.hpp"
 #include "CfgVehicles.hpp"
 #include "RscTitles.hpp"
-#include "Dialog.hpp"
