+++ addons/medical_feedback/functions/fnc_handleHUDIndicators.sqf
@@ -17,7 +17,6 @@
  */
 params [["_dropAllIndicators", false]];
 
-private _hudContainer = uiNamespace getVariable [QGVAR(HUDIndicatorsContainer), controlNull];
 private _indicatorSlots = [
     uiNamespace getVariable [QGVAR(stateIndicator1), controlNull],
     uiNamespace getVariable [QGVAR(stateIndicator2), controlNull],
@@ -29,7 +28,6 @@ if (_dropAllIndicators) exitWith {
     {
         _x ctrlSetText "";
     } forEach _indicatorSlots;
-    _hudContainer ctrlCommit 0;
 };
 
 // --- Tourniquets
@@ -64,5 +62,3 @@ private _icons = [
 {
     _x ctrlSetText (_icons param [_forEachIndex, ""]);
 } forEach _indicatorSlots;
-
-_hudContainer ctrlCommit 0;
